Hotfix: Stats index page should not use integers to store download counts (#7545)

+++ src/NuGetGallery/ViewModels/StatisticsFact.cs
@@ -7,13 +7,13 @@ namespace NuGetGallery
 {
     public class StatisticsFact
     {
-        public StatisticsFact(IDictionary<string, string> dimensions, int amount)
+        public StatisticsFact(IDictionary<string, string> dimensions, long amount)
         {
             Dimensions = new Dictionary<string, string>(dimensions);
             Amount = amount;
         }
 
         public IDictionary<string, string> Dimensions { get; private set; }
-        public int Amount { get; private set; }
+        public long Amount { get; private set; }
     }
 }
