+++ src/NuGetGallery/ViewModels/StatisticsNuGetUsageItem.cs
@@ -6,6 +6,6 @@ namespace NuGetGallery
     public class StatisticsNuGetUsageItem
     {
         public string Version { get; set; }
-        public int Downloads { get; set; }
+        public long Downloads { get; set; }
     }
 }
